Rework logic around running probers
This part of the code has always been very messy. Move the logic around
creating a probe out of the scraper package into its own packages and
add a Prober interface that the different probers have to implement in
order to be able to run them.
