+++ internal/prober/logger/logger.go
@@ -1,0 +1,5 @@
+package logger
+
+type Logger interface {
+	Log(keyvals ...interface{}) error
+}
